Fix a bug while closing TCP workers
Closing reader before writer causes the reader not to stop. This
should be some winsock issue: closing a dup socket descriptor does not
cause the reader to be interrupted. By closing the writer worker first,
the writer socket is closed cleanly and when reader socket is closed,
the reader thread is interrupted.

This change also introduces new logging events and error checks.