Bug fixing: ensure serial buffers are actually purged
If we just purge after connection is established, we will clear the data reached until that instant. But if the device is still sending bytes they will be queued in the buffer. 

The fix consists in waiting 10 millis for the Arduino to reset. Once reset is in progress, we can be sure no more bytes will be sent until reset is completed so purge is sure at this point as soon as we do not wait too much for the board to send its first message.